--- tasks/mpi/ermolaev_v_allreduce/src/ops_mpi.cpp
@@ -1,2 +1,0 @@
-// Copyright 2023 Nesterov Alexander
-#include "mpi/ermolaev_v_allreduce/include/ops_mpi.hpp"
