+++ tasks/mpi/ermolaev_v_allreduce_library/include/ops_mpi.hpp
@@ -1,0 +1,4 @@
+// Copyright 2023 Nesterov Alexander
+#pragma once
+
+#include "mpi/ermolaev_v_allreduce_my/include/ops_mpi.hpp"
