+++ tasks/mpi/ermolaev_v_allreduce_library/include/test_funcs.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "mpi/ermolaev_v_allreduce_my/include/test_funcs.hpp"
