+++ tasks/mpi/ermolaev_v_allreduce_library/src/ops_mpi.cpp
@@ -1,0 +1,2 @@
+// Copyright 2023 Nesterov Alexander
+#include "mpi/ermolaev_v_allreduce_library/include/ops_mpi.hpp"
