+++ tasks/mpi/ermolaev_v_allreduce_my/include/allreduce.hpp
@@ -2,7 +2,7 @@
 
 #include <boost/mpi/communicator.hpp>
 
-#include "mpi/ermolaev_v_allreduce/include/shared_ptr_array.hpp"
+#include "mpi/ermolaev_v_allreduce_my/include/shared_ptr_array.hpp"
 
 namespace ermolaev_v_allreduce_mpi {
