+++ tasks/mpi/ermolaev_v_allreduce_my/include/shared_ptr_array.hpp
