+++ tasks/mpi/ermolaev_v_allreduce_my/src/ops_mpi.cpp
@@ -1,0 +1,2 @@
+// Copyright 2023 Nesterov Alexander
+#include "mpi/ermolaev_v_allreduce_my/include/ops_mpi.hpp"
