--- tasks/seq/ermolaev_v_allreduce/src/ops_mpi.cpp
@@ -1,2 +1,0 @@
-// Copyright 2023 Nesterov Alexander
-#include "seq/ermolaev_v_allreduce/include/ops_seq.hpp"
